add analyzer, add dynamic imports and direct imports to reduce first load size
