+++ package.json
@@ -15,11 +15,13 @@
     "format": "prettier --write .",
     "start": "next start",
     "twconfig": "tailwind-config-viewer -o --port 3600",
-    "test": "jest"
+    "test": "jest",
+    "analyze": "ANALYZE=true yarn build"
   },
   "dependencies": {
     "@elastic/elasticsearch": "8.1.0",
     "@next/env": "12.0.8",
+    "@next/bundle-analyzer": "12.1.4",
     "autoprefixer": "10.4.2",
     "axios": "0.24.0",
     "classnames": "2.3.1",
