+++ pages/404.jsx
@@ -4,7 +4,8 @@ import { useRouter } from 'next/router'
 import cls from 'classnames'
 import { longTextClass } from '@/components/Typo'
 import { i18n } from '@/next-i18next.config'
-import { map, omit } from 'lodash'
+import map from 'lodash/map'
+import omit from 'lodash/omit'
 import { getMenus } from '@/lib/ssr-api'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
