+++ pages/500.jsx
@@ -1,7 +1,8 @@
 import Head from 'next/head'
 import { useRouter } from 'next/router'
 import cls from 'classnames'
-import { map, omit } from 'lodash'
+import map from 'lodash/map'
+import omit from 'lodash/omit'
 import { BlankLayout, FALLBACK_TITLE } from '@/components/layout/Layout'
 import { longTextClass } from '@/components/Typo'
 import { i18n } from '@/next-i18next.config'
