+++ pages/[[...slug]].jsx
@@ -70,7 +70,7 @@ export async function getStaticProps(context) {
   const { params, locale } = context
   params.slug = params.slug || ['/']
   const path = params.slug.join('/')
-  const isNodePath = /node/.test(params.slug[0])
+  // const isNodePath = /node/.test(params.slug[0])
 
   const type = await getResourceTypeFromContext({
     locale,
@@ -108,20 +108,20 @@ export async function getStaticProps(context) {
     return NOT_FOUND
   }
 
-  if (isNodePath) {
-    if (node.path?.alias) {
-      return {
-        redirect: {
-          permanent: false,
-          destination: `/${locale}/${node.path.alias}`,
-        },
-      }
-    } else {
-      console.warn(
-        `Warning: request to direct node ${path} blocked. 404 returned `
-      )
-    }
-  }
+  // if (isNodePath) {
+  //   if (node.path?.alias) {
+  //     return {
+  //       redirect: {
+  //         permanent: false,
+  //         destination: `/${locale}/${node.path.alias}`,
+  //       },
+  //     }
+  //   } else {
+  //     console.warn(
+  //       `Warning: request to direct node ${path} blocked. 404 returned `
+  //     )
+  //   }
+  // }
 
   let fiNode = null // Must be JSON compatible
