+++ src/components/layout/CityMenu.jsx
@@ -7,7 +7,7 @@ import { useTranslation } from 'next-i18next'
 import useSWR from 'swr'
 import { useRouter } from 'next/router'
 import { getMunicipalities } from '@/lib/client-api'
-import { times } from 'lodash'
+import times from 'lodash/times'
 
 const Loader = () =>
   times(12, (i) => (
