+++ src/components/messages/Messages.jsx
@@ -2,7 +2,7 @@ import { useAtom } from 'jotai'
 import { useTranslation } from 'next-i18next'
 import { useAtomValue } from 'jotai/utils'
 import { nodeIdAtom, shownMessagesAtom } from '@/src/store'
-import { keys } from 'lodash'
+import keys from 'lodash/keys'
 import LanguageMessageCard from '@/components/messages/LanguageMessageCard'
 import MessageCard from './MessageCard'
 import useSWR from 'swr'
